+++ CHANGELOG.md
@@ -2,13 +2,14 @@
 Blip is the web front end for YourLoops system.
 It is based on Tidepool Blip 1.27.
 
-## 3.0.2-beta17 - 2022-11-10
+## 3.0.2-beta23 - 2022-11-17
 ### Added
 - YLP-1679 Allow caregivers to remove patients from their direct share
 - YLP-1850 Patient search should comply to INS standard
 ### Engineering use
 - YLP-1624 Fix error messages when inviting a patient already in the team
 - YLP-1680 getPatientData Api method has a wrong parameter
+- YLP-1805 Push error to the backend
 - YLP-1873 Fix patient profile update not visible in dashboard
 - YLP-1874 Fix PDF report generation
 - YLP-1892 Creation of an integration test for the team form
@@ -41,6 +42,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-1767 Create package to migrate viz components
 - YLP-1797 Make it possible to see what role the users are logged in with matomo
 - YLP-1834 Migrate time in range stats
+- YLP-1915 Migrate reservoir tooltip
 - Add a route /version to display blip version
 
 ### Fixed